Refactor Actor into an abstract class and implement UUID.
- Converted `Actor` interface into an abstract class.
- Moved common fields (`uuid`, `particlePool`, `tile`, `tileCoord`, `health`, `dead`, `status`, `inventory`) into `Actor`.
- Implemented core methods from `HasPosition`, `HasHealth`, and `HasInventory` in the `Actor` base class.
- Updated `CardPlayer` and `Structure` to extend `Actor` and remove redundant code.
- Updated `ProcChainTest` to match the new `Actor` structure.
- Every `Actor` now receives a unique, immutable `java.util.UUID` on creation.
