+++ nomad-realms/src/main/java/nomadrealms/context/game/actor/Actor.java
@@ -7,6 +7,7 @@ import static nomadrealms.context.game.actor.status.StatusEffect.INVINCIBLE;
 import engine.visuals.constraint.box.ConstraintPair;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.UUID;
 import nomadrealms.context.game.GameState;
 import nomadrealms.context.game.actor.status.Status;
 import nomadrealms.context.game.actor.types.HasHealth;
@@ -17,10 +18,13 @@ import nomadrealms.context.game.card.effect.SpawnParticlesEffect;
 import nomadrealms.context.game.card.query.actor.SelfQuery;
 import nomadrealms.context.game.event.InputEvent;
 import nomadrealms.context.game.event.Target;
+import nomadrealms.context.game.item.Inventory;
 import nomadrealms.context.game.world.World;
 import nomadrealms.context.game.world.map.area.Tile;
+import nomadrealms.context.game.world.map.area.coordinate.TileCoordinate;
 import nomadrealms.event.game.effect.EffectContext;
 import nomadrealms.render.Renderable;
+import nomadrealms.render.particle.NullParticlePool;
 import nomadrealms.render.particle.ParticlePool;
 import nomadrealms.render.particle.spawner.BasicParticleSpawner;
 
@@ -30,29 +34,65 @@ import nomadrealms.render.particle.spawner.BasicParticleSpawner;
  *
  * @author Lunkle
  */
-public interface Actor extends HasPosition, HasHealth, HasInventory, Target, Renderable {
+public abstract class Actor implements HasPosition, HasHealth, HasInventory, Target, Renderable {
 
-	String name();
+	private transient final UUID uuid = UUID.randomUUID();
 
-	default List<Action> actions() {
+	private transient ParticlePool particlePool = new NullParticlePool();
+
+	protected TileCoordinate tileCoord;
+	protected transient Tile tile;
+
+	protected int health;
+	protected boolean dead;
+
+	protected final Status status = new Status();
+	protected final Inventory inventory = new Inventory(this);
+
+	protected Actor() {
+	}
+
+	public UUID uuid() {
+		return uuid;
+	}
+
+	public abstract String name();
+
+	public List<Action> actions() {
 		return new ArrayList<>();
 	}
 
-	default void update(GameState state) {
+	public void update(GameState state) {
 	}
 
-	default List<InputEvent> retrieveNextPlays() {
+	public List<InputEvent> retrieveNextPlays() {
 		return new ArrayList<>();
 	}
-
-	boolean dead();
-
-	void dead(boolean dead);
-
-	Status status();
 
 	@Override
-	default void damage(int damage) {
+	public int health() {
+		return health;
+	}
+
+	@Override
+	public void health(int health) {
+		this.health = health;
+	}
+
+	public boolean dead() {
+		return dead;
+	}
+
+	public void dead(boolean dead) {
+		this.dead = dead;
+	}
+
+	public Status status() {
+		return status;
+	}
+
+	@Override
+	public void damage(int damage) {
 		if (damage > 0 && status().count(INVINCIBLE) > 0) {
 			status().remove(INVINCIBLE, 1);
 			particlePool().addParticles(new SpawnParticlesEffect(
@@ -68,26 +108,48 @@ public interface Actor extends HasPosition, HasHealth, HasInventory, Target, Ren
 		HasHealth.super.damage(damage);
 	}
 
-	void particlePool(ParticlePool particlePool);
+	public void particlePool(ParticlePool particlePool) {
+		this.particlePool = particlePool;
+	}
 
-	ParticlePool particlePool();
+	public ParticlePool particlePool() {
+		return particlePool;
+	}
+
+	@Override
+	public Inventory inventory() {
+		return inventory;
+	}
+
+	@Override
+	public Tile tile() {
+		return tile;
+	}
+
+	@Override
+	public void tile(Tile tile) {
+		this.tile = tile;
+		this.tileCoord = tile.coord();
+	}
 
 	/**
 	 * @return Whether this actor has the capability of restocking cards when they run out, or if they will simply die once
 	 * they are out of cards.
 	 */
-	default boolean shouldRestock() {
+	public boolean shouldRestock() {
 		return true;
 	}
 
 	/**
 	 * purely done for the sake of adding references to optimize other algorithms
 	 */
-	default void reindex(World world) {
+	public void reindex(World world) {
+		tile = world.getTile(tileCoord);
+		inventory.reindex(this);
 	}
 
 	@Override
-	default boolean move(Tile target) {
+	public boolean move(Tile target) {
 		if (target.actor() != null) {
 			return false;
 		}
